requirements.txt
deepgram-sdk
openai
pyaudio
click
